new commit

+++ 0x08-recursion/100-is_palindrome.c
@@ -15,7 +15,7 @@
 int is_pal(char *s, int i)
 {
 
-	if (i == strlen(s) / 2)
+	if (i == (int)strlen(s) / 2)
 		return (1);
 	if (*(s + i) == *(s + strlen(s) - i - 1))
 		return (is_pal(s, i + 1));
